Move codegen_llvm::common::ty_fn_sig into rustc::ty::Instance.
